#485 Show alias info message on article detail page
