+++ next/public/locales/sk/translation.json
@@ -1,5 +1,6 @@
 {
-  "aliasSection.aliasInfoMessage": "Túto stránku nájdete aj na",
+  "aliasInfoMessage.message,page": "",
+  "aliasInfoMessage.message.article": "Tento článok nájdete aj na",
   "aria.close": "Zavrieť",
   "articlePage.shareblock.buttonText": "Zdieľať článok",
   "articlePage.shareblock.text": "Páčil sa vám článok?",
@@ -35,13 +36,13 @@
   "gallery.aria.openGallery": "Otvoriť galériu obrázkov",
   "galleryModal.aria.closeGallery": "Zavrieť galériu obrázkov",
   "gallerySlider.aria.descriptionSlider": "Posúvač",
+  "globalSearch.enterSearchQuery": "Zadajte hľadaný výraz",
+  "globalSearch.noResults": "Žiadne výsledky",
   "globalSearch.searchOption.allResults": "Všetky výsledky",
   "globalSearch.searchOption.article": "Článok",
   "globalSearch.searchOption.articles": "Články",
   "globalSearch.searchOption.document": "Dokument",
   "globalSearch.searchOption.documents": "Dokumenty",
-  "globalSearch.enterSearchQuery": "Zadajte hľadaný výraz",
-  "globalSearch.noResults": "Žiadne výsledky",
   "globalSearch.searchOption.page": "Stránka",
   "globalSearch.searchOption.pages": "Stránky",
   "globalSearch.searchOptions": "Možnosti vyhľadávania",
